+++ dfm_tools/modelbuilder.py
@@ -9,14 +9,13 @@ Since the functions in this script contain hardcoded parameters, it is not expos
 """
 
 import os
-import xarray as xr
 import pandas as pd
 import dfm_tools as dfmt
 import hydrolib.core.dflowfm as hcdfm
 import datetime as dt
-import glob
 from hydrolib.core.dimr.models import DIMR, FMComponent, Start
 import warnings
+from dfm_tools.hydrolib_helpers import get_ncbnd_construct
 
 
 def cmems_nc_to_bc(ext_bnd, list_quantities, tstart, tstop, file_pli, dir_pattern, dir_output, refdate_str):
@@ -50,46 +49,79 @@ def cmems_nc_to_bc(ext_bnd, list_quantities, tstart, tstop, file_pli, dir_patter
     
     return ext_bnd
 
+
+def preprocess_ini_cmems_to_nc(**kwargs):
+    raise DeprecationWarning("`dfmt.preprocess_ini_cmems_to_nc()` was deprecated, use `cmems_nc_to_ini()` instead")
+
+
+def cmems_nc_to_ini(ext_old, dir_output, list_quantities, tstart, dir_pattern, conversion_dict=None):
     
-def preprocess_ini_cmems_to_nc(ext_old, tstart, dir_data, dir_out):
+    if conversion_dict is None:
+        conversion_dict = dfmt.get_conversion_dict()
     
-    file_nc_list_so = glob.glob(f'{dir_data}\\cmems_so_*.nc')
-    file_nc_list_thetao = glob.glob(f'{dir_data}\\cmems_thetao_*.nc')
-    file_nc_list = file_nc_list_so + file_nc_list_thetao
-    
-    print(f'opening {len(file_nc_list)} datasets')
-    data_xr = xr.open_mfdataset(file_nc_list)
-    
-    # fill nans
-    # start with lat/lon to avoid values from shallow coastal areas in deep layers
-    # first interpolate nans to get smooth filling of e.g. islands, this cannot fill nans at the edge of the dataset
-    data_xr = data_xr.interpolate_na(dim='latitude').interpolate_na(dim='longitude')
-    
-    # then use bfill/ffill to fill nans at the edge for lat/lon/depth
-    data_xr = data_xr.ffill(dim='latitude').bfill(dim='latitude')
-    data_xr = data_xr.ffill(dim='longitude').bfill(dim='longitude')
-    data_xr = data_xr.ffill(dim='depth').bfill(dim='depth')
-    
-    tSimStart = pd.Timestamp(tstart)
-    data_xr_ontime = data_xr.sel(time=slice(tSimStart-dt.timedelta(days=1),tSimStart+dt.timedelta(days=1)))
-    
-    print('writing file')
-    outFile = os.path.join(dir_out,f'InitialField_{tSimStart.strftime("%Y-%m-%d_%H-%M-%S")}.nc')
-    data_xr_ontime.to_netcdf(outFile,format="NETCDF4_CLASSIC") #TODO: why the format setting?
-    
-    #append forcings to ext
-    # 3D initialsalinity/initialtemperature fields are silently ignored, initial 3D conditions are only possible via nudging 1st timestep via quantity=nudge_salinity_temperature
-    forcing_saltem = hcdfm.ExtOldForcing(quantity='nudge_salinity_temperature',
-                                         filename=outFile,
-                                         filetype=hcdfm.ExtOldFileType.NetCDFGridData,
-                                         method=hcdfm.ExtOldMethod.InterpolateTimeAndSpaceSaveWeights, #3
-                                         operand=hcdfm.Operand.override, #O
-                                         )
-    ext_old.forcing.append(forcing_saltem)
+    tstart_pd = pd.Timestamp(tstart)
+    tstart_str = tstart_pd.strftime("%Y-%m-%d_%H-%M-%S")
+    tstart_round, tstop_round = dfmt.round_timestamp_to_outer_noon(tstart,tstart)
+    for quan_bnd in list_quantities:
+        
+        if quan_bnd in ["temperaturebnd","uxuyadvectionvelocitybnd"]:
+            # silently skipped, temperature is handled with salinity, uxuy not supported
+            continue
+        elif quan_bnd=="salinitybnd":
+            # 3D initialsalinity/initialtemperature fields are silently ignored
+            # initial 3D conditions are only possible via nudging 1st timestep via quantity=nudge_salinity_temperature
+            data_xr = dfmt.open_dataset_extra(dir_pattern=dir_pattern, quantity=["salinitybnd","temperaturebnd"],
+                                              tstart=tstart_round, tstop=tstop_round,
+                                              conversion_dict=conversion_dict)
+            data_xr = data_xr.rename_vars({"salinitybnd":"so", "temperaturebnd":"thetao"})
+            quantity = "nudge_salinity_temperature"
+            varname = None
+        else:
+            data_xr = dfmt.open_dataset_extra(dir_pattern=dir_pattern, quantity=quan_bnd,
+                                              tstart=tstart_round, tstop=tstop_round,
+                                              conversion_dict=conversion_dict)
+            quantity = f'initial{quan_bnd.replace("bnd","")}'
+            varname = quantity
+            data_xr = data_xr.rename_vars({quan_bnd:quantity})
+        
+        # open_dataset_extra converted depths from positive down to positive up, including update of the "positive" attribute
+        # TODO: this correctly updated attr negatively impacts model results when using netcdf inifields, so we revert it here
+        # https://issuetracker.deltares.nl/browse/UNST-7455
+        ncbnd_construct = get_ncbnd_construct()
+        varn_depth = ncbnd_construct['varn_depth']
+        if varn_depth in data_xr.coords:
+            data_xr[varn_depth].attrs['positive'] = 'down'
+        
+        # subset two times. interp to tstart would be the proper way to do it, 
+        # but FM needs two timesteps for nudge_salinity_temperature and initial waq vars
+        data_xr = data_xr.sel(time=slice(tstart_round, tstop_round))
+        
+        # fill nans, start with lat/lon to avoid values from shallow coastal areas in deep layers
+        # first interpolate nans to get smooth filling of e.g. islands, this cannot fill nans at the edge of the dataset
+        data_xr = data_xr.interpolate_na(dim='latitude').interpolate_na(dim='longitude')
+        
+        # then use bfill/ffill to fill nans at the edge for lat/lon/depth
+        data_xr = data_xr.ffill(dim='latitude').bfill(dim='latitude')
+        data_xr = data_xr.ffill(dim='longitude').bfill(dim='longitude')
+        data_xr = data_xr.ffill(dim='depth').bfill(dim='depth')
+        
+
+        print('writing file')
+        file_output = os.path.join(dir_output,f"{quantity}_{tstart_str}.nc")
+        data_xr.to_netcdf(file_output)
+        
+        #append forcings to ext
+        forcing_saltem = hcdfm.ExtOldForcing(quantity=quantity,
+                                             varname=varname,
+                                             filename=file_output,
+                                             filetype=hcdfm.ExtOldFileType.NetCDFGridData,
+                                             method=hcdfm.ExtOldMethod.InterpolateTimeAndSpaceSaveWeights, #3
+                                             operand=hcdfm.Operand.override) #O
+        ext_old.forcing.append(forcing_saltem)
     
     return ext_old
-    
-    
+
+
 def preprocess_merge_meteofiles_era5(ext_old, varkey_list, dir_data, dir_output, time_slice):
 
     if isinstance(varkey_list[0], list):